♻️ [refactor] #5 CH2 키오스크 과제 필드에 final 추가하여 불변성 보장
